Don't call .reconfigure twice, refactor
